% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofi_categories.R
\name{add_broad_ofi_categories}
\alias{add_broad_ofi_categories}
\title{Add broad OFI categories to data}
\usage{
add_broad_ofi_categories(data)
}
\arguments{
\item{data}{A data frame containing the 'Problemomrade_.FMP' column}
}
\value{
The input data frame with a new column 'ofi.categories.broad'
}
\description{
This function adds broad Opportunities for Improvement (OFI) categories to the input data frame.
}
