% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofi_categories.R
\name{add_detailed_ofi_categories}
\alias{add_detailed_ofi_categories}
\title{Add detailed OFI categories to data}
\usage{
add_detailed_ofi_categories(data)
}
\arguments{
\item{data}{A data frame containing the 'Problemomrade_.FMP' column}
}
\value{
The input data frame with a new column 'ofi.categories.detailed'
}
\description{
This function adds detailed Opportunities for Improvement (OFI) categories to the input data frame.
}
