% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ofi_categories.R
\name{add_ofi_categories}
\alias{add_ofi_categories}
\title{Add OFI categories to data}
\usage{
add_ofi_categories(data)
}
\arguments{
\item{data}{A data frame containing the 'Problemomrade_.FMP' column}
}
\value{
The input data frame with two new columns: 'ofi.categories.detailed' and 'ofi.categories.broad'
}
\description{
This function adds both detailed and broad Opportunities for Improvement (OFI)
categories to the input data frame.
}
