% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tqip_cohorts.R, R/create_cohorts.R
\name{add_tqip_cohorts}
\alias{add_tqip_cohorts}
\title{Add TQIP cohorts to a dataset}
\usage{
add_tqip_cohorts(dataset)

add_tqip_cohorts(dataset)
}
\arguments{
\item{dataset}{A data frame containing AIS codes as columnes with names containing AIS as created by the rofi package. No default.}
}
\value{
The original data frame with an added collumn "cohort" containing assigned cohorts

The original data frame with the added column "cohort" containing the added cohorts
}
\description{
This functions adds the TQIP cohorts Blunt multisystem without TBI, Blunt multistystem with TBI,
Isolated severe TBI and Severe penetrating to a dataset.

This functions adds the TQIP cohorts Blunt multisystem without TBI, Blunt multistystem with TBI,
Isolated severe TBI and Severe penetrating to a dataset.
}
\section{Prerequisites}{

Need to gave gone through the clean_all_predictors() and clean_audit_filters() function. In order to account for intubation status and GCS.
}

\section{How to use}{

\code{new.dataset <- add_tqip_cohorts(combined.dataset) #apply function to combined.dataset and save in a new.dataset}
}

