% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ofi_categories.R
\name{create_broad_ofi_categories}
\alias{create_broad_ofi_categories}
\title{Create broad OFI categories}
\usage{
create_broad_ofi_categories(data)
}
\arguments{
\item{data}{A data frame containing the 'Problemomrade_.FMP' column}
}
\value{
A character vector of broad OFI categories
}
\description{
This function creates broad Opportunities for Improvement (OFI) categories
based on the detailed categories created by create_detailed_ofi_categories().
}
