% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofi_categories.R
\name{create_detailed_ofi_categories}
\alias{create_detailed_ofi_categories}
\title{Create detailed OFI categories}
\usage{
create_detailed_ofi_categories(data)
}
\arguments{
\item{data}{A data frame containing the 'Problemomrade_.FMP' column}
}
\value{
A character vector of detailed OFI categories
}
\description{
This function creates detailed Opportunities for Improvement (OFI) categories
based on the 'Problemomrade_.FMP' column in the input data.
}
