% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intub.R
\name{create_intub}
\alias{create_intub}
\title{Create Intubation Variable}
\usage{
create_intub(data)
}
\arguments{
\item{data}{A data frame containing the data from which the
intubation variable will be created.}
}
\value{
A vector indicating the intubation status for each patient.
}
\description{
This function creates a new variable indicating whether a patient
was intubated, regardless of whether the intubation was performed
prehospital or in the emergency department. The result is a vector
containing the intubation status for each patient.
}
\examples{
data$intub <- create_intub(data)

}
