% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_variables_to_be_included.R
\name{get_default_variables_to_be_included}
\alias{get_default_variables_to_be_included}
\title{Get Default Variables to Include}
\usage{
get_default_variables_to_be_included()
}
\value{
A character vector containing the names of default variables.
}
\description{
This function returns a character vector containing the names of
default variables that are included in the prepare_data
function. These variables cover various aspects such as patient
information, injury details, time-related data, and outcomes.
}
\examples{
default.variables <- get_default_variables_to_be_included()

}
