% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Data}
\usage{
import_data(
  user,
  db.name = "opportunities_for_improvement",
  table.names = c("swetrau", "fmp", "atgarder", "problem"),
  password = keyring::key_get(db.name, user)
)
}
\arguments{
\item{user}{Character. Your username to connect to the database. No default.}

\item{db.name}{Character. The name of the database. Defaults to "opportunities_for_improvement".}

\item{table.names}{Character. The name(s) of the table(s) in the database. Defaults to c("swetrau", "fmp", "atgarder", "problem").}

\item{password}{Character. Your password to connect to the database. Defaults to keyring::key_get(db.name, user).}
}
\description{
Function to import data from the local MariaDB database.
}
