% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_data}
\alias{label_data}
\title{Label Data Function}
\usage{
label_data(data)
}
\arguments{
\item{data}{A data frame containing the data to be labeled and converted. No default.}
}
\value{
A labeled data frame with updated factor levels and column names.
}
\description{
This function takes a data frame as input and performs various labeling and factor conversion operations on its columns. The labeled data is returned with new factor levels and column names.
}
