% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fmp_with_problem.R
\name{merge_fmp_with_problem}
\alias{merge_fmp_with_problem}
\title{Merge FMP and problem}
\usage{
merge_fmp_with_problem(fmp.data, problem.data)
}
\arguments{
\item{fmp.data}{A data.frame. The FMP data. No default.}

\item{problem.data}{A data.frame. The data describing the opportunities for improvement. No default.}
}
\description{
This function merges the File Maker Pro database from 2017 to 2021
with a column describing the opportunities for improvment. These
have to be merged because they were distributed to us as separate
files.
}
