% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_old_review_data_with_swetrau.R
\name{merge_old_review_data_with_swetrau}
\alias{merge_old_review_data_with_swetrau}
\title{Merge old review data with SweTrau}
\usage{
merge_old_review_data_with_swetrau(swetrau.data, old.review.data)
}
\arguments{
\item{swetrau.data}{A data.frame. The SweTrau data. No default.}

\item{old.review.data}{A data.frame. The review data from the period 2014-2017. No default.}
}
\description{
This function merges review data from the period 2014 to 2017 with SweTrau. During this period the review process was not formalised and the data structure different, so it takes quite a but of manual data manipulation to get it to work.
}
