% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data Function}
\usage{
prepare_data(
  data,
  variables.to.be.included = get_default_variables_to_be_included()
)
}
\arguments{
\item{data}{A data frame containing the raw data to be prepared.}

\item{variables.to.be.included}{A character vector specifying the
names of variables to be included in the prepared data. Default
is to include the default set of variables returned by
\code{\link{get_default_variables_to_be_included}}.}
}
\value{
A prepared data frame with added variables, subsetted
columns, and labeled data.
}
\description{
This function takes a data frame and a vector of variable names to
be included and performs several steps to prepare the data for
analysis. It adds computed variables, subsets the data to include
only specified variables, labels the data, and returns the prepared
data frame.
}
\examples{
prepared.data <- prepare_data(data)

}
