% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ofi.R
\name{create_ofi}
\alias{create_ofi}
\title{Create Opportunities For Improvement Variable}
\usage{
create_ofi(
  data,
  quality.review.done = "VK_avslutad",
  problem.area = "Problemomrade_.FMP",
  mortality.review.done = "tra_DodsfallsanalysGenomford",
  preventable.death = "Fr1.14"
)
}
\arguments{
\item{data}{A data.frame. The data needed to create the
opportunities for improvements variable. Has to include columns
with tne names specified in the arguments quality.review.done,
problem.area, mortality.review.done, and preventable.death.  No
default.}

\item{quality.review.done}{Character. The name of the quality
review done variable, which indicates if the quality review has
been completed. Defaults to "VK_avslutad".}

\item{problem.area}{Character. The name of the problem area
variable, which indicates what, if any, problem area that were
identified. Defaults to "Problemomrade_.FMP".}

\item{mortality.review.done}{Character. The name of the mortatlity
review done variable, which indicates if the mortality review
has been completed. Defaults to "tra_DodsfallsanalysGenomford".}

\item{preventable.death}{Character. The name of the preventable
death variable, which indicates if a death was prevantable,
potentially proventable, and not preventable. Defaults to
"Fr1.14".}
}
\description{
Create opportunities for improvement variable, which is either yes
or no
}
\details{
Yes: The case was flagged and reviewed in a meeting and the
consensus was that there were opportunities for improvement
(variable Problemomrade_.FMP), or that the patient died and the
death was determined as preventable or potentially preventable
(variable Fr1.14 2 or 3).

No: The consensus was that there were no opportunities for
improvement, or the nurses in the initial review did not send the
case for review because everything was okay (variable VK_avslutad).

NA: The case was never selected for review or the patient died but
whether the death was preventable is registered as unknown (Fr1.14
999)
}
