% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Data}
\usage{
import_data(
  user = Sys.getenv("DB_USERNAME"),
  password = Sys.getenv("DB_PASSWORD"),
  db.name = "opportunities_for_improvement",
  table.names = c("swetrau", "fmp", "atgarder", "problem", "kvalgranskning2014.2017"),
  silent = FALSE
)
}
\arguments{
\item{user}{Character. Your username to connect to the
database. Defaults to Sys.getenv("DB_USERNAME").}

\item{password}{Character. Your password to connect to the
database. Defaults to Sys.getenv("DB_PASSWORD").}

\item{db.name}{Character. The name of the database. Defaults to
"opportunities_for_improvement".}

\item{table.names}{Character. The name(s) of the table(s) in the
database. Defaults to c("swetrau", "fmp", "atgarder",
"problem", "kvalgranskning2014.2017").}

\item{silent}{Logical. If FALSE a message is displayed with the
names of the imported datasets. Defaults to FALSE.}
}
\value{
A list of data.frames.
}
\description{
Function to import data from the local MariaDB database. This
assumes that your username and password are available as the
environment variables DB_USERNAME and DB_PASSWORD. \code{rofi} uses the
package \code{dotenv} to access environment variables and load any .env
file in the project directory. We therefore recommed storing the
username and password in an .env file in the project directory, but
make sure this file is ignored by git so that you don't put it in
version control by mistake. All this is taken care of by using
\code{rofi::init()}.
}
