% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge data}
\usage{
merge_data(datasets)
}
\arguments{
\item{datasets}{A list of data.frames. Has to include the datasets
"swetrau", "fmp", "atgarder", and "problem". The dataset
"swetrau" has to include the variables
"DateTime_ArrivalAtHospital", "PersonIdentity", and
"TempIdentity". The datasets "fmp" and "problem" have to
include the variables "Ankomst_te", "Personnummer", and
"Reservnummer".}
}
\description{
Merges the different opportunities for improvement datasets into a
single data.frame.
}
